[Refactor] Extract protocol KnobMapping to KnobMapping.swift
